DPO-Internlm2_chat_1_8B/requirements.txt
transformers
torch
sentencepiece
einops
mdtex2html
llama-cpp-python==0.2.69
gradio==4.28.3
huggingface_hub==0.22.2